refactor(errors): separate tool error runtime
