+++ src/code-ops.ts
@@ -10,7 +10,8 @@ import type {
   EditCodeRule,
   EditCodeRuleObject,
 } from "./code-contract";
-import { createToolError, type EditCodeRecoveryKind, TOOL_ERROR_CODES, type ToolRecovery } from "./error-primitives";
+import { type EditCodeRecoveryKind, TOOL_ERROR_CODES, type ToolRecovery } from "./error-primitives";
+import { createToolError } from "./tool-error";
 import { createDiff, displayPathForDiff, ensurePathWithinAllowedRoots, IGNORED_DIRS } from "./tool-utils";
 
 function editCodeRecovery(path: string, kind: EditCodeRecoveryKind): ToolRecovery {
