+++ src/file-ops.ts
@@ -1,6 +1,7 @@
 import { mkdir, readFile, unlink, writeFile } from "node:fs/promises";
 import { dirname, join } from "node:path";
-import { createToolError, type EditFileRecoveryKind, TOOL_ERROR_CODES, type ToolRecovery } from "./error-primitives";
+import { type EditFileRecoveryKind, TOOL_ERROR_CODES, type ToolRecovery } from "./error-primitives";
+import { createToolError } from "./tool-error";
 
 /** Owner-only read/write. Use for files containing secrets or sensitive metadata. */
 export const PRIVATE_FILE_MODE = 0o600;
