+++ src/tool-execution.ts
@@ -1,6 +1,7 @@
 import { invariant } from "./assert";
-import { ERROR_KINDS, LIFECYCLE_ERROR_CODES, ToolError, type ToolRecovery } from "./error-primitives";
+import { ERROR_KINDS, LIFECYCLE_ERROR_CODES, type ToolRecovery } from "./error-primitives";
 import { createId } from "./short-id";
+import { ToolError } from "./tool-error";
 import { recordCall, runGuards, type SessionContext } from "./tool-guards";
 
 function withTimeout<T>(task: () => Promise<T>, timeoutMs: number, toolId: string): Promise<T> {
